test(marktree): move test functions into marktree_test fixture
In order ot run unittests with a release build, we need the test
functions to be accessible when NDEBUG is defined. Moving the functions
into the test fixture ensures they are available and only available for
use by the unit tests.

Since the tests and code use some common static functions, prefix those
with mt_ and remove the static declaration.